Move unit tests for setting applicant1 solicitor answers to submit event
As the code to generate the applicant1 solicitor answers has been moved into
the SolicitorSubmitJointApplication event, I've moved the unit tests over
to cover it
